+++ pkg/proxy/runner/bounded_frequency_runner.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package async
+package runner
 
 import (
 	"fmt"
